% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atletas}
\alias{atletas}
\title{Jogos olímpicos de verão: 1896 até 2016}
\format{
Uma tibble com 186.069 linhas e 12 variáveis. Os nomes das variáveis estão em português, mas os valores estão em inglês.
\describe{
\item{id}{Identificador único do atleta (pois há atletas diferentes com o mesmo nome)}
\item{nome}{Nome do atleta}
\item{sexo}{'F' ou 'M'}
\item{idade}{Idade do atleta no momento da participação}
\item{altura}{Altura em cm}
\item{peso}{Peso em kg}
\item{pais}{Sigla do comitê olímpico do país (veja a tibble \code{siglas})}
\item{ano}{Ano da participação}
\item{cidade}{Cidade onde aconteceram os jogos}
\item{esporte}{Esporte}
\item{evento}{Evento (dentro do esporte)}
\item{medalha}{Se o atleta ganhou medalha na participação, um string 'Gold', 'Silver', ou 'Bronze'; caso contrário, \code{NA}}
}
}
\source{
\url{https://www.kaggle.com/heesoo37/120-years-of-olympic-history-athletes-and-results}, modificado por fnaufel
}
\usage{
atletas
}
\description{
Dados sobre participações de atletas nos jogos olímpicos de verão. Cada atleta pode aparecer em diversas linhas, uma linha para cada participação.
}
\keyword{datasets}
