% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{copiar_enunciado}
\alias{copiar_enunciado}
\title{copiar_enunciado}
\usage{
copiar_enunciado(caminho = ".")
}
\arguments{
\item{caminho}{diretório ou arquivo de destino (string)}
}
\value{
(invisível) nome do arquivo de destino
}
\description{
Copia arquivo .Rmd com o enunciado da atividade para um destino
da sua escolha (default é o diretório atual). Gera erro se o arquivo
de destino já existe.
}
