% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siglas}
\alias{siglas}
\title{Siglas dos comitês olímpicos de diversos países}
\format{
Uma tibble com 230 linhas e 3 variáveis. Os nomes das variáveis estão em português, mas os valores estão em inglês.
\describe{
\item{sigla}{sigla em maiúsculas}
\item{regiao}{país ou região}
\item{obs}{observações ou \code{NA}}
}
}
\source{
\url{https://www.kaggle.com/heesoo37/120-years-of-olympic-history-athletes-and-results}, modificado por fnaufel
}
\usage{
siglas
}
\description{
Siglas dos comitês olímpicos de diversos países
}
\keyword{datasets}
